moved `env_value_to_bool` to env_tools
